// src/app/pages/endereco-estabelecimento/model/endereco-estabelecimento.ts
export class EnderecoEstabelecimento {
    restaurantID: number;
    addressID: number;
    zipCode: string;
    street: string
    sublocality: string;
    number: string;
    city: string;
    state: string;
    country: string;
    latitude: string;
    longitude: string;
    observations: string;
}